requirements.txt
fire
Pillow
pillow_heif